Resolves eslint error.
