+++ package.json
@@ -27,7 +27,7 @@
     "koa-router": "^7.4.0",
     "lodash": "^4.17.11",
     "mysql2": "^1.6.4",
-    "reflect-metadata": "^0.1.12",
+    "reflect-metadata": "^0.1.13",
     "routing-controllers": "^0.7.7",
     "type-graphql": "^0.16.0",
     "typeorm": "^0.2.9"
