+++ src/data/entities/Place.ts
@@ -1,19 +1,7 @@
 import { Entity, Column } from 'typeorm'
 import { BaseEntry } from './BaseEntry'
-import { Field, ObjectType, registerEnumType } from 'type-graphql'
-
-export enum PlaceType {
-  country = 'country', // eslint-disable-line no-unused-vars
-  state = 'state', // eslint-disable-line no-unused-vars
-  region = 'region', // eslint-disable-line no-unused-vars
-  city = 'city', // eslint-disable-line no-unused-vars
-  district = 'district', // eslint-disable-line no-unused-vars
-}
-
-// required for type-graphql
-registerEnumType(PlaceType, {
-  name: 'PlaceType'
-})
+import { PlaceType } from './PlaceType'
+import { Field, ObjectType } from 'type-graphql'
 
 /**
  * Entity class mapping rows in 'place' table.
