+++ src/type-graphql/resolvers/PlaceResolvers.ts
@@ -1,18 +1,17 @@
 import { Place } from '../../data/entities/'
 import { PlaceController } from '../../controllers'
-import { ListArgs } from './types/ListArgs'
-import { Resolver, Query, Arg, Args } from 'type-graphql'
+import { Resolver, Query, Arg } from 'type-graphql'
 
 @Resolver(Place)
 export class PlaceResolver {
-
   @Query(returns => Place)
   place (@Arg('id') id: number) {
     return PlaceController.getInstance().getOne(id)
   }
 
   @Query(returns => [Place])
-  places (@Args() { first, offset }: ListArgs) {
+  places (@Arg('first', { nullable: true }) first: number,
+          @Arg('offset', { nullable: true }) offset: number) {
     return PlaceController.getInstance().getAll(first, offset)
   }
 }
